requirements.txt
flask
flask-sqlalchemy
flask-login
flask-migrate
python-dotenv
pandas
numpy
prophet
tensorflow
scikit-learn
langchain
langchain-community
langchain-experimental
matplotlib
seaborn
plotly
reportlab
apscheduler
gunicorn
werkzeug
yfinance
pypdf
python-jose
Passlib
bcrypt
tabulate
